# Allow activation function for Diagonal