Following up on #1911 , this PR allows for an activation function argument to be passed to `Flux.Diagonal` and reroutes the `LayerNorm` activation function to `Flux.Diagonal` now. This reduces memory and time taken for `LayerNorm` with an activation function.

Before this PR (Julia master, Flux master):

```julia
julia> l = LayerNorm(768, relu);

julia> x = rand(Float32, 768, 128, 2);

julia> @benchmark $(l)($x)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  193.291 μs …   4.904 ms  ┊ GC (min … max):  0.00% … 91.62%
 Time  (median):     310.625 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   385.946 μs ± 415.163 μs  ┊ GC (mean ± σ):  19.75% ± 16.07%

  ▂▃█▇▂                                                         ▁
  █████▆▅▅▄▃▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▃▄▄▆▆▆▇▅▆▇▆▇▇▆▆▆▅▅▆▅▅▅▆▅▅▁▅ █
  193 μs        Histogram: log(frequency) by time       2.75 ms <

 Memory estimate: 2.25 MiB, allocs estimate: 18.
```

After this PR:
```julia
julia> @benchmark $(l)($x)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  168.333 μs …   1.870 ms  ┊ GC (min … max):  0.00% … 84.42%
 Time  (median):     245.667 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   281.375 μs ± 209.594 μs  ┊ GC (mean ± σ):  13.78% ± 14.90%

  ▄▃▄█▇▃                                                        ▁
  ███████▅▅▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▄▅▅▆▆▇▇██▇▇▇▇ █
  168 μs        Histogram: log(frequency) by time       1.43 ms <

 Memory estimate: 1.50 MiB, allocs estimate: 16.
```

PS: there are some formatting changes that I just made because they caught my eye, but if they're jarring I can always remove them