@theabhirath 2022-04-02
There was no type restriction on the input for `Flux.Diagonal` earlier - there's a test to that effect checking if it works with just an integer as an input (which it did earlier). The fast activation path now is incompatible with that though, so is this something I should manually work around or is an `AbstractArray` restriction actually better for the layer given the general usecase?