@CarloLucibello 2022-04-02
I think restricting to abstract array is fine and this is a good moment to introduce a breaking change