@ToucheSir 2022-04-02
Let's do it, I highly doubt anyone was passing non-array inputs anyhow.